+++ DragonSpark.Application.Environment.Development/QueryContext.cs
@@ -1,0 +1,5 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace DragonSpark.Application.Environment.Development;
+
+public readonly record struct QueryContext(Type Type, QueryTrackingBehavior Behavior);
